% Auto-generated: do not edit by hand
\name{dgtc_nestedMyNestedComponent}

\alias{dgtc_nestedMyNestedComponent}

\title{MyNestedComponent component}

\description{
MyNestedComponent description
}

\usage{
dgtc_nestedMyNestedComponent(id=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The id of the component}

\item{value}{Character. The value to display}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

